+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ### New Features
  - add computability with `user_input` refactoring from SelfHelp v7.0.0 
 ### Bugfix
- - check if there is selected survey in the style before trying to save
+ - check if there is a selected survey in the style before trying to save
+ - do not try to save the survey in CMS mode
 
 # v1.3.5
 ### Bugfix